requirements.txt
fastapi
uvicorn
tensorflow
numpy
pillow
pytest
python-multipart
